+++ advanced_examples/column_and_cell_rendering/mock_data.json
@@ -1,0 +1,109 @@
+{
+    "stocks": [
+      {
+        "ticker": "AAPL",
+        "name": "Apple Inc.",
+        "price": 150.5,
+        "marketCap": 2500000000,
+        "change": 1.25
+      },
+      {
+        "ticker": "GOOGL",
+        "name": "Alphabet Inc.",
+        "price": 2800.75,
+        "marketCap": 1900000000,
+        "change": -0.75
+      },
+      {
+        "ticker": "MSFT",
+        "name": "Microsoft Corporation",
+        "price": 300.25,
+        "marketCap": 220000000,
+        "change": 0.98
+      },
+      {
+        "ticker": "AMZN",
+        "name": "Amazon.com Inc.",
+        "price": 3500.55,
+        "marketCap": 1700000000,
+        "change": 2.15
+      },
+      {
+        "ticker": "TSLA",
+        "name": "Tesla Inc.",
+        "price": 800.45,
+        "marketCap": 800000000,
+        "change": -1.10
+      },
+      {
+        "ticker": "FB",
+        "name": "Meta Platforms Inc.",
+        "price": 340.75,
+        "marketCap": 900000000,
+        "change": 0.50
+      },
+      {
+        "ticker": "NFLX",
+        "name": "Netflix Inc.",
+        "price": 590.65,
+        "marketCap": 250000000,
+        "change": 1.75
+      },
+      {
+        "ticker": "NVDA",
+        "name": "NVIDIA Corporation",
+        "price": 220.15,
+        "marketCap": 600000000,
+        "change": 0.85
+      },
+      {
+        "ticker": "BABA",
+        "name": "Alibaba Group Holding Limited",
+        "price": 180.25,
+        "marketCap": 500000000,
+        "change": -0.95
+      },
+      {
+        "ticker": "V",
+        "name": "Visa Inc.",
+        "price": 230.45,
+        "marketCap": 450000000,
+        "change": 1.05
+      },
+      {
+        "ticker": "JPM",
+        "name": "JPMorgan Chase & Co.",
+        "price": 160.35,
+        "marketCap": 400000000,
+        "change": 0.65
+      },
+      {
+        "ticker": "DIS",
+        "name": "The Walt Disney Company",
+        "price": 180.75,
+        "marketCap": 300000000,
+        "change": -0.45
+      },
+      {
+        "ticker": "PYPL",
+        "name": "PayPal Holdings Inc.",
+        "price": 250.85,
+        "marketCap": 350000000,
+        "change": 1.25
+      },
+      {
+        "ticker": "ADBE",
+        "name": "Adobe Inc.",
+        "price": 650.95,
+        "marketCap": 550000000,
+        "change": 0.95
+      },
+      {
+        "ticker": "INTC",
+        "name": "Intel Corporation",
+        "price": 55.75,
+        "marketCap": 200000000,
+        "change": -0.25
+      }
+    ]
+  }
