+++ advanced_examples/grouping_widgets/widgets.json
@@ -1,0 +1,44 @@
+{
+  "historical_chains_table": {
+    "name": "Chains Historical table example",
+    "description": "Get historical TVL of a chain using Defi LLama",
+    "category": "crypto",
+    "searchCategory": "crypto",
+    "endpoint": "historical_chains",
+    "gridData": {
+      "w": 20,
+      "h": 9
+    },
+    "params": [
+      {
+        "paramName": "chain",
+        "description": "Select a chain to get historical TVL",
+        "value": "Ethereum",
+        "label": "Chain",
+        "type": "endpoint",
+        "optionsEndpoint": "get_chains_list"
+      }
+    ]
+  },
+  "historical_chains_table_advanced_dropdown": {
+    "name": "Chains Historical table example with advanced dropdown",
+    "description": "Get historical TVL of a chain using Defi LLama",
+    "category": "crypto",
+    "searchCategory": "crypto",
+    "endpoint": "historical_chains",
+    "gridData": {
+      "w": 20,
+      "h": 9
+    },
+    "params": [
+      {
+        "paramName": "chain",
+        "description": "Select a chain to get historical TVL",
+        "value": "Ethereum",
+        "label": "Chain",
+        "type": "endpoint",
+        "optionsEndpoint": "get_chains_list_advanced"
+      }
+    ]
+  }
+}
