+++ database_connection_examples/clickhouse_python/widgets.json
