+++ database_connection_examples/elasticsearch_python/widgets.json
