+++ database_connection_examples/mindsdb_python/widgets.json
