+++ database_connection_examples/snowflake_connector_python/.env.example
