+++ database_connection_examples/snowflake_connector_python/utils.py
