+++ database_connection_examples/supabase_python/widgets.json
