+++ widget_examples/chart_widget/plotly_templates.py
