+++ widget_examples/chart_widget/widgets.json
@@ -1,0 +1,48 @@
+{
+  "chains_plotly": {
+    "name": "Chains chart example",
+    "description": "Get current TVL of all chains and plot it with Plotly!",
+    "category": "crypto",
+    "type": "chart",
+    "endpoint": "chains",
+    "gridData": {
+      "w": 20,
+      "h": 9
+    }
+  },
+  "chains_table": {
+    "name": "Chains Chart example",
+    "description": "A built in chart widget example",
+    "category": "Crypto",
+    "endpoint": "chains_table",
+    "gridData": {
+      "w": 20,
+      "h": 9
+    },
+    "data": {
+      "table": {
+        "chartView": {
+            "enabled": true,
+            "chartType": "bar"
+        },
+        "showAll": false,
+        "columnsDefs": [
+            {
+                "headerName": "Name",
+                "field": "name",
+                "chartDataType": "category"
+            },
+            {
+                "headerName": "Total Value Locked",
+                "field": "tvl"
+            },
+            {
+                "headerName": "Token Symbol",
+                "field": "tokenSymbol",
+                "chartDataType": "category"
+            }
+        ]
+      }
+    }
+  }
+}
