+++ widget_examples/requirements.txt
@@ -3,3 +3,4 @@ uvicorn==0.22.0
 pandas==2.0.1
 plotly==5.15.0
 requests==2.31.0
+textwrap
