+++ widget_examples/table_widget/main.py
@@ -1,0 +1,52 @@
+import json
+from pathlib import Path
+import pandas as pd
+import requests
+from fastapi import FastAPI
+from fastapi.middleware.cors import CORSMiddleware
+from fastapi.responses import JSONResponse
+
+app = FastAPI()
+
+origins = [
+    "https://pro.openbb.co",
+    "https://excel.openbb.co",
+    "http://localhost:1420"
+]
+
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=origins,
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
+
+ROOT_PATH = Path(__file__).parent.resolve()
+
+@app.get("/")
+def read_root():
+    return {"Info": "Full example for OpenBB Custom Backend"}
+
+
+@app.get("/widgets.json")
+def get_widgets():
+    """Widgets configuration file for the OpenBB Custom Backend"""
+    return JSONResponse(
+        content=json.load((Path(__file__).parent.resolve() / "widgets.json").open())
+    )
+
+
+@app.get("/chains_table")
+def chains_table():
+    """Get current TVL of all chains using Defi LLama"""
+    params = {}
+    response = requests.get("https://api.llama.fi/v2/chains", params=params)
+
+    if response.status_code == 200:
+        return response.json()
+
+    print(f"Request error {response.status_code}: {response.text}")
+    return JSONResponse(
+        content={"error": response.text}, status_code=response.status_code
+    )
