--- widgets.json
@@ -1,71 +1,0 @@
-{
-    "custom_widget": {
-        "name": "Custom Widget Example",
-        "description": "A widget to demonstrate custom configuration",
-        "endpoint": "custom-endpoint",
-        "data": {
-            "dataKey": "customDataKey",
-            "table": {
-                "enableCharts": true,
-                "showAll": true,
-                "chartView": {
-                    "enabled": true,
-                    "chartType": "column"
-                },
-                "columnsDefs": [
-                    {
-                        "field": "column1",
-                        "headerName": "Column 1",
-                        "chartDataType": "category",
-                        "cellDataType": "text",
-                        "formatterFn": "none",
-                        "renderFn": "titleCase",
-                        "width": 100,
-                        "maxWidth": 200,
-                        "minWidth": 50,
-                        "hide": false,
-                        "pinned": "left"
-                    },
-                    {
-                        "field": "column2",
-                        "headerName": "Column 2",
-                        "chartDataType": "series",
-                        "cellDataType": "number",
-                        "formatterFn": "int",
-                        "renderFn": "greenRed",
-                        "width": 150
-                    }
-                ]
-            }
-        },
-        "params": [
-            {
-                "type": "date",
-                "paramName": "startDate",
-                "value": "2024-01-01",
-                "label": "Start Date",
-                "show": true,
-                "description": "The start date for the data",
-                "options": [],
-                "multiSelect": false
-            },
-            {
-                "type": "text",
-                "paramName": "ticker",
-                "value": "AAPL",
-                "label": "Ticker",
-                "show": true,
-                "description": "Stock ticker symbol",
-                "options": [],
-                "multiSelect": true
-            }
-        ],
-        "source": [
-            "API",
-            "Database"
-        ],
-        "defaultViz": "table",
-        "staleTime": 300000,
-        "refreshInterval": 60000
-    }
-}
